+++ src/main/java/com/billit/user_service/account/dto/response/AccountInvestResponse.java
@@ -6,11 +6,12 @@ import lombok.Getter;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.util.UUID;
 
 @Getter
 @Builder
 public class AccountInvestResponse {
-    private Long id;
+    private Integer id;
     private String bankName;
     private String accountNumber;
     private String accountHolder;
